[dart2wasm] Check const-ness of instructions added to const exprs
A constant expression in Wasm is an instruction sequence that consists
of only the constant instructions.

Currently if I generate a non-constant instruction in a constant
expression I get an error during Wasm compilation. I then have to find
where and how I generate the instruction, possibly by using `--watch
<instruction offset>` passed to dart2wasm.

With this change we check in an assertion that an instruction added to
an instruction sequence for a constant expression can only be a constant
expression.

This helps with debugging as I notice the error earlier, and get a stack
trace of the compiler function that generates the invalid instruction.

Changes:

- Add `bool get isConstant => false` member to `Instruction`.

  A method instead of field to avoid adding one more word to
  `Instruction` objects.

- Override the member in constant instructions.

- When creating an instructions builder, mark the builder as "constant",
  with an optional named constructor argument that defaults to `false`.

- When adding an instruction to an instructions builder, check that if
  the instruction block is for a constant expression, the instruction is
  constant.

Spec:
https://webassembly.github.io/gc/core/valid/instructions.html#constant-expressions

Change-Id: I59454f0634ab419a62f9455c84aa9ce631db8b63
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/391440
Commit-Queue: Ömer Ağacan <omersa@google.com>
Reviewed-by: Slava Egorov <vegorov@google.com>