[analyzer] Rename whyNotPromotedList -> whyNotPromotedArguments
The new name better reflects the purpose of the argument, which is to
keep track of "why not promoted" information for the arguments of the
invocation.

In follow-up CLs I'll be doing more plumbing of "why not promoted"
information, and this rename will make it clearer to distinguish the
new plumbing from `whyNotPromotedArguments`.

Change-Id: Ie4d618c211221fc0ea8e3a1c79f24a3423663f9e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/391566
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>