[dart2wasm] Fix dispatch table class ID loop bounds.
The loop iterates over the class IDs up to (but not including) maxConcreteClassId. But maxConcreteClassId is the actual ID of the last concrete class so the loop is ending 1 early.

From some testing, WasmTable seems to always be the last class. This class doesn't have any subclasses so all the calls to it are likely devirtualized anyway skipping the dispatch table. In fact the way the class IDs are assigned, the last class should always be a leaf concrete class. I believe that's why this hasn't caused a detectable bug.

Change-Id: I7003dfd015c2ad13a4a042ae86af5e4f54242452
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/391862
Commit-Queue: Nate Biggs <natebiggs@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>