[flow] Part 2 Issue 3658 - Forward IfNullSet let expressions in the CFE.
This is a small change that allows forwarding let expressions for IfNullSets in the CFE.

Since we now store the promotion information for a written expression in an if-statement, the CFE builds a let expression and it's not recognized by flow analysis. To fix this, we add a forward so flow analysis recognizes that it's the same expression that it's trying to find.

language/inference_update_4/assignment_promotion_in_if_statement_test passes from this change.

Bug: https://github.com/dart-lang/language/issues/3658
Change-Id: I4ec99de036498869b0d9c5c652ca748d9f5e570a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/391861
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>