[dart2wasm] Move chunked JSON parser state to a base class
Moved the state in the mixin to a base class to reduce duplication in
generated code.

Members that deal with the state that are not overridden and don't use
overridden members are also moved to the base class, as they can't
introduce polymorphism.

Change-Id: Ie71ecd617d01409cdaae1ec3b9bd039d72266c88
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/394440
Reviewed-by: Slava Egorov <vegorov@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>