[analysis_server] Allow cancellation tokens to carry a code/reason
This means when we cancel requests ourselves (such as for completion/refactor when another request begins) we can provide more specific codes/reasons.

This change only uses the reason, but when the scheduler is cancelling requests (because of a `/didChange`) it could use the `ContentModified` code preserving existing behaviour where those cancellations are currently handled by the session change.

Change-Id: I48eea55aff08b15a539a1c2c6031598adb4dd069
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/394444
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Commit-Queue: Keerti Parthasarathy <keertip@google.com>