analyzer: Use Registry.ruleRegistry instead of a rule provider in LinterRuleOptionsValidator.
The theme of this change is that the Registry.ruleRegistry is always
the source of truth for what lint rules are registered, even in tests.
Before this change, there was a mechanism for passing a "rule
provider" around different classes and functions, which is a function
that returns a list of lint rules. This mechanism only existed for one
test, `options_rule_validator_test.dart`.

So this change instead makes that test use Registry.ruleRegistry, and
removes all of the rule provider mechanisms. All in all:

* Remove the LintRuleProvider type alias.
* Remove the LinterRuleOptionsValidator constructor's `provider`
  parameter and corresponding field.
* Remove the `provider` parameter from `analyzeAnalysisOptions()`.
* Remove the `provider` parameter from `OptionsFileValidator()`.
* In AbstractAnalysisOptionsTest, remove the `provider` parameter from
  `assertErrorsInCode`.
* Rename `_validatePluginsOption` to `_validateLegacyPluginsOption`
  while I'm there.
* In order to register the test LintRule classes, they need a proper
  `lintCode` property. So create a super class for them, simplifying
  each test LintRule class.


Change-Id: Ic153bba006fe0fe2a94abb08ac55723ffb2c356e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/393907
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>