[analyzer] Add DartObject.toRecordValue()
The `DartObject` class in the analyzer represents a known constant
value. It has methods that convert that representation into a more
specific one (e.g. `toListValue()` that returns a list of
`DartObject`).

While constant records can be represented as `DartObject`, there's no
method to directly extract the positional and named fields from that
instance. Users could extract the shape from the record type and then
call `getField`, but that is an error-prone process.
So, this adds `toRecordValue()`, which returns a list for positional
and a map for named parameters when called on a `DartObject`
representing a record.

TEST=pkg/analyzer/test/src/dart/constant/value_test.dart

Change-Id: I56f3e877ea78ab0f3ccfb7ac73ffda80ee4fe846
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/394860
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>