[mini_types] Decouple NamedFunctionParameter from NamedType.
Previously, NamedFunctionParameter was derived from NamedType. This
avoided a tiny amount of code duplication, but the benefit was really
tiny. Decoupling the two classes simplifies the implementation of a
follow-up CL that I have in progress, which will allow the mini_types
framework to represent function types.

Change-Id: I29bff36ab76bbf344cf7204f2503b006a88a4cf8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/395422
Reviewed-by: Kallen Tu <kallentu@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>