[cfe] Extract the lowered wildcard name for initializing formals when inferring field type.
When we introduced lowering wildcard variable names in https://dart-review.googlesource.com/c/sdk/+/373621, we missed the fact that we also need to extract the name in `finalizeInitializingFormal` so we can infer the correct type.

This caused all formals to be initialized with the `dynamic` type instead.

This fix makes sure that we're inferring the correct type for these wildcard variable initializing formals.

Bug: https://github.com/dart-lang/sdk/issues/57089
Change-Id: I380bf960082dc395a55b8cbed4753b2f009af2dd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/395343
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>