Elements. Add PropertyAccessorElement2 and PropertyAccessorFragment.
I found a number of places where we need to access `variable3`, so
it might be convenient to have a common interface that provides it.

Still, there are many places where we want to know that this is
specifically `GetterElement`, to separate interfaces for them are
also useful.

Change-Id: Ib671dc3bdb9a04a022fd74921f8ccedcdba79643
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/395701
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>